- Reaction roles rewritten completely. They now support multiple exclusivity groups per message and level requirements. However they can only be added one by one - Bot now support much higher XP values for global and server levels

+++ src/WizBot/Db/Models/DiscordUser.cs
@@ -14,7 +14,7 @@ public class DiscordUser : DbEntity
     public ClubInfo Club { get; set; }
     public bool IsClubAdmin { get; set; }
 
-    public int TotalXp { get; set; }
+    public long TotalXp { get; set; }
     public DateTime LastLevelUp { get; set; } = DateTime.UtcNow;
     public DateTime LastXpGain { get; set; } = DateTime.MinValue;
     public XpNotificationLocation NotifyOnLevelUp { get; set; }
