+++ src/WizBot/Db/Models/UserXpStats.cs
@@ -5,8 +5,8 @@ public class UserXpStats : DbEntity
 {
     public ulong UserId { get; set; }
     public ulong GuildId { get; set; }
-    public int Xp { get; set; }
-    public int AwardedXp { get; set; }
+    public long Xp { get; set; }
+    public long AwardedXp { get; set; }
     public XpNotificationLocation NotifyOnLevelUp { get; set; }
     public DateTime LastLevelUp { get; set; } = DateTime.UtcNow;
 }
